remove obsolete propAmong
